Añado ViewHolder

+++ Recu.sql
@@ -2,7 +2,7 @@ CREATE TABLE tbLibros (
 UUID_Libro VARCHAR2(100) PRIMARY KEY,
 tituloLibro VARCHAR2(75) NOT NULL,
 autorLibro VARCHAR2(50) NOT NULL,
-añoPublicacion NUMBER(4,0) NOT NULL,
+añoPublicacion INT NOT NULL,
 estadoLibro VARCHAR2(15) NOT NULL CHECK (estadoLibro IN ('nuevo', 'usado', 'seminuevo')),
 ISBM VARCHAR2(30) NOT NULL,
 generoLibro VARCHAR2(50) NOT NULL,
@@ -19,5 +19,4 @@ clave VARCHAR2(100) NOT NULL
 );
 
 select * from tbUsuarios;
-
 SELECT * FROM tbUsuarios WHERE nombreUsuario = ? AND apellidoUsuario = ? AND correoElectronico = ? AND clave = ?
